lazy load analyze components
